Remove duplicated packages (blaming IntelliJ)
