+++ subprojects/build-adapter-xcode/src/test/java/dev/nokee/buildadapter/xcode/AllXCWorkspaceLocationsValueSourceTests.java
@@ -13,10 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.buildadapter.xcode.dev.nokee.buildadapter.xcode;
+package dev.nokee.buildadapter.xcode;
 
-import dev.nokee.buildadapter.xcode.internal.plugins.AllXCProjectLocationsValueSource;
-import dev.nokee.buildadapter.xcode.internal.plugins.XCProjectLocator;
+import dev.nokee.buildadapter.xcode.internal.plugins.AllXCWorkspaceLocationsValueSource;
+import dev.nokee.buildadapter.xcode.internal.plugins.XCWorkspaceLocator;
+import dev.nokee.platform.xcode.EmptyXCWorkspace;
+import dev.nokee.xcode.XCWorkspace;
+import dev.nokee.xcode.XCWorkspaceReference;
+import lombok.EqualsAndHashCode;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -27,10 +31,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.nio.file.Path;
 
-import static dev.nokee.buildadapter.xcode.dev.nokee.buildadapter.xcode.TestProjectReference.project;
 import static dev.nokee.internal.testing.FileSystemMatchers.aFile;
 import static dev.nokee.internal.testing.FileSystemMatchers.withAbsolutePath;
 import static dev.nokee.internal.testing.util.ProjectTestUtils.objectFactory;
+import static dev.nokee.xcode.XCWorkspaceReference.of;
 import static java.util.Arrays.asList;
 import static java.util.Collections.emptyList;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -40,17 +44,17 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.mockito.hamcrest.MockitoHamcrest.argThat;
 
 @ExtendWith(MockitoExtension.class)
-class AllXCProjectLocationsValueSourceTests {
+class AllXCWorkspaceLocationsValueSourceTests {
 	@TempDir Path testDirectory;
-	@Mock XCProjectLocator locator;
-	AllXCProjectLocationsValueSource.Parameters parameters;
-	AllXCProjectLocationsValueSource subject;
+	@Mock XCWorkspaceLocator locator;
+	AllXCWorkspaceLocationsValueSource.Parameters parameters;
+	AllXCWorkspaceLocationsValueSource subject;
 
 	@BeforeEach
 	void createSubject() {
-		parameters = objectFactory().newInstance(AllXCProjectLocationsValueSource.Parameters.class);
+		parameters = objectFactory().newInstance(AllXCWorkspaceLocationsValueSource.Parameters.class);
 		parameters.getSearchDirectory().set(testDirectory.toFile());
-		subject = new AllXCProjectLocationsValueSource(locator) {
+		subject = new AllXCWorkspaceLocationsValueSource(locator) {
 			@Override
 			public Parameters getParameters() {
 				return parameters;
@@ -59,24 +63,47 @@ class AllXCProjectLocationsValueSourceTests {
 	}
 
 	@Test
-	void returnsProjectReferenceAsFoundByLocator() {
-		Mockito.when(locator.findProjects(usingSearchDirectory())).thenReturn(asList(
-			project("A.xcodeproj"), project("C.xcodeproj"),
-			project("B.xcodeproj")));
+	void returnsWorkspaceReferenceAsFoundByLocator() {
+		Mockito.when(locator.findWorkspaces(usingSearchDirectory())).thenReturn(asList(
+			workspace("A.xcworkspace"), workspace("C.xcworkspace"),
+			workspace("B.xcworkspace")));
 
 		assertThat(subject.obtain(), contains(
-			project("A.xcodeproj"), project("C.xcodeproj"),
-			project("B.xcodeproj")));
+			workspace("A.xcworkspace"), workspace("C.xcworkspace"),
+			workspace("B.xcworkspace")));
 	}
 
 	@Test
-	void returnsEmptyIterableWhenLocatorDoesNotFoundProjects() {
-		Mockito.when(locator.findProjects(usingSearchDirectory())).thenReturn(emptyList());
+	void returnsEmptyIterableWhenLocatorDoesNotFoundWorkspaces() {
+		Mockito.when(locator.findWorkspaces(usingSearchDirectory())).thenReturn(emptyList());
 
 		assertThat(subject.obtain(), emptyIterable());
 	}
 
 	private Path usingSearchDirectory() {
 		return argThat(aFile(withAbsolutePath(equalTo(testDirectory.toAbsolutePath().toString()))));
+	}
+
+	private static XCWorkspaceReference workspace(String name) {
+		return new TestWorkspaceReference(name);
+	}
+
+	@EqualsAndHashCode
+	private static final class TestWorkspaceReference implements XCWorkspaceReference {
+		private final String name;
+
+		private TestWorkspaceReference(String name) {
+			this.name = name;
+		}
+
+		@Override
+		public Path getLocation() {
+			throw new UnsupportedOperationException();
+		}
+
+		@Override
+		public XCWorkspace load() {
+			throw new UnsupportedOperationException();
+		}
 	}
 }
