+++ subprojects/build-adapter-xcode/src/test/java/dev/nokee/buildadapter/xcode/XcodeProjectPathRuleTests.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package dev.nokee.buildadapter.xcode.dev.nokee.buildadapter.xcode;
+package dev.nokee.buildadapter.xcode;
 
 import dev.nokee.buildadapter.xcode.internal.GradleProjectPathService;
 import dev.nokee.buildadapter.xcode.internal.components.GradleProjectPathComponent;
